% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collect_tatts_sales.R
\name{collect_tatts_sales}
\alias{collect_tatts_sales}
\title{Collect Data}
\usage{
collect_tatts_sales(url, sale_name = "NULL")
}
\arguments{
\item{\strong{url}}{the url of the sale}

\item{\strong{sale_name}}{the name of the sale}
}
\description{
This function allows users to collect data from a specific tattersalls sale
that might not be included in the package.
}
\details{
The format of the data displayed on the pages appears to be uniform
between the different sales, however the data is collected via regular
expressions which potentially cause headaches for you the user, apologies. If
you do come across a sale that this function cannot collect for let me know
and I will attempt to rectify that.  A list of available sales to collect data
for can be found here \url{http://db.tattersalls.com/4DCGI/Entry/MenuResults},
but please check that the sale you want isn't in this package, find a list
of sales here \link{tattssales}
}

