% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{tatts_2010}
\alias{tatts_2010}
\title{Tattersalls Sales in 2010}
\format{7796 rows, 12 columns}
\description{
Dataset of 12 Tattersalls sales in 2010
}
\section{tatts_2010}{

\itemize{
     \item \strong{dec_mares_2010} December Mares 2010 Sale
     \item \strong{dec_foals_2010} December Foals 2010 Sale
     \item \strong{dec_yearlings_2010} December Yearlings 2010 Sale
     \item \strong{autumn_hit_2010} Autumn Horses in Training 2010 Sale
     \item \strong{breeders_flat_2010} Breeders Flat 2010 Sale
     \item \strong{oct_bk3_2010} October Book 3 2010 Sale
     \item \strong{oct_bk2_2010} October Book 2 2010 Sale
     \item \strong{oct_bk1_2010} October Book 1 2010 Sale
     \item \strong{july_2010} July 2010 Sale
     \item \strong{guineas_2010} Breeze Up 2010 Sale
     \item \strong{craven_2010} Craven 2010 Sale
     \item \strong{feb_2010} February 2010 Sale
}
}
\keyword{dataset}

