% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBloodstock.R
\docType{data}
\name{tatts_2011}
\alias{tatts_2011}
\title{Tattersalls Sales in 2011}
\format{6946 rows, 12 columns}
\description{
Dataset of 12 Tattersalls sales in 2011
}
\section{tatts_2011}{

\itemize{
     \item dec_mares_2011 December Mares 2011 Sale
     \item dec_foals_2011 December Foals 2011 Sale
     \item dec_yearlings_2011 December Yearlings 2011 Sale
     \item autumn_hit_2011 Autumn Horses in Training 2011 Sale
     \item breeders_flat_2011 Breeders Flat 2011 Sale
     \item oct_bk3_2011 October Book 3 2011 Sale
     \item oct_bk2_2011 October Book 2 2011 Sale
     \item oct_bk1_2011 October Book 1 2011 Sale
     \item july_2011 July 2011 Sale
     \item breeze_up_2011 Breeze Up 2011 Sale
     \item craven_2011 Craven 2011 Sale
     \item feb_2011 February 2011 Sale
}
}
\keyword{dataset}

