% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{tatts_2013}
\alias{tatts_2013}
\title{Tattersalls Sales in 2013}
\format{7192 rows, 12 columns}
\description{
Dataset of 14 Tattersalls sales in 2013
}
\section{Sales}{

\itemize{
     \item \strong{dec_mares_2013} December Mares 2013 Sale
     \item \strong{dec_foals_2013} December Foals 2013 Sale
     \item \strong{dec_yearlings_2013} December Yearlings 2013 Sale
     \item \strong{autumn_hit_2013} Autumn Horses in Training 2013 Sale
     \item \strong{oct_bk4_2013} October Book 4 2013 Sale
     \item \strong{oct_bk3_2013} October Book 3 2013 Sale
     \item \strong{oct_bk2_2013} October Book 2 2013 Sale
     \item \strong{oct_bk1_2013} October Book 1 2013 Sale
     \item \strong{july_2013} July 2013 Sale
     \item \strong{guineas_hit_2013} Guineas Horses in Training 2013 Sale
     \item \strong{guineas_2013} Guineas 2013 Sale
     \item \strong{craven_2013} Craven 2013 Sale
     \item \strong{feb_2013} February 2013 Sale
}
}
\keyword{dataset}

