% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{tatts_2014}
\alias{tatts_2014}
\title{Tattersalls Sales in 2014}
\format{1469 rows, 12 columns}
\description{
Dataset of 5 Tattersalls sales in 2014
}
\section{tatts_2014}{

\itemize{
     \item \strong{july_2014} July 2014 Sale
     \item \strong{guineas_hit_2014} Guineas Horses in Training 2014 Sale
     \item \strong{guineas_2014} Guineas 2014 Sale
     \item \strong{craven_2014} Craven 2014 Sale
     \item \strong{feb_2014} February 2014 Sale
}
}
\keyword{dataset}

