% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rBloodstock.R
\docType{data}
\name{tatts_2015}
\alias{tatts_2015}
\title{Tattersalls sales in 2015}
\format{307 rows, 12 columns}
\description{
Dataset of Tattersalls sales in 2015
}
\section{tatts_2015}{

\itemize{
     \item feb_2015 February 2015 sale
}
}
\keyword{dataset}

