% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rBloodstock.R
\docType{package}
\name{rBloodstock}
\alias{rBloodstock}
\alias{rBloodstock-package}
\title{rBloodstock}
\source{
\url{http://db.tattersalls.com/4DCGI/Entry/Index}
}
\description{
A growing collection of Thoroughbred Horse sales datasets
}
\section{Datasets}{

\itemize{
     \item \link{tatts_2015} 1537 rows, consisting of four sales (more will be
     added)
     \item \link{tatts_2014} 7429 rows, made up of 13 individual sales
     \item \link{tatts_2013} 7192 rows, made up of 13 individual sales
     \item \link{tatts_2012} 7101 rows, made up of 12 individual sales
     \item \link{tatts_2011} 6946 rows, made up of 12 individual sales
     \item \link{tatts_2010} 7796 rows, made up of 12 individual sales
}
}

\section{Tatts Variables}{

\itemize{
     \item year year of sale
     \item sale sale name (often same year on year)
     \item sale_name concatenation of \emph{sale} and \emph{year}
     \item lot_no lot number
     \item horse horse (plenty of unnamed)
     \item sex horses' sex
     \item color horses' color
     \item sire horses' sire
     \item dam horses' dam
     \item consignor seller
     \item purchaser buyer
     \item price in guineas
}
}

\section{sex}{

\itemize{
     \item c colt
     \item f filly
     \item g gelding
     \item h horse
     \item m mare
}
}

\section{color}{

\itemize{
     \item b bay
     \item br brown
     \item ch chesnut
     \item gr grey
     \item ro roan
}
}
\author{
Thomas Heslop
}

