% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rBloodstock}
\alias{rBloodstock}
\alias{rBloodstock-package}
\title{rBloodstock}
\source{
\url{http://db.tattersalls.com/4DCGI/Entry/Index}
}
\description{
A growing collection of Thoroughbred Horse sales datasets
}
\section{Datasets}{

\itemize{
     \item \link{tatts_2014} 1469 rows, made up of 5 individual sales (more
     will be added)
     \item \link{tatts_2013} 8280 rows, made up of 14 individual sales
     \item \link{tatts_2012} 7739 rows, made up of 14 individual sales
     \item \link{tatts_2011} 7488 rows, made up of 13 individual sales
     \item \link{tatts_2010} 8346 rows, made up of 13 individual sales
}
}

\section{Tatts Variables}{

\itemize{
     \item \strong{year} year of sale
     \item \strong{sale} sale name (often same year on year)
     \item \strong{sale_name} concatenation of \emph{sale} and \emph{year}
     \item \strong{lot_no} lot number
     \item \strong{horse} horse (plenty of unnamed)
     \item \strong{sex} horses' sex
     \item \strong{color} horses' color
     \item \strong{sire} horses' sire
     \item \strong{dam} horses' dam
     \item \strong{consignor} seller
     \item \strong{purchaser} buyer
     \item \strong{price} in guineas
}
}

\section{sex}{

\itemize{
     \item \strong{c} colt
     \item \strong{f} filly
     \item \strong{g} gelding
     \item \strong{h} horse
     \item \strong{m} mare
}
}

\section{color}{

\itemize{
     \item \strong{b} bay
     \item \strong{br} brown
     \item \strong{ch} chesnut
     \item \strong{gr} grey
     \item \strong{ro} roan
}
}
\author{
Thomas Heslop
}

