% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rBloodstock.R
\docType{data}
\name{tatts_2010}
\alias{tatts_2010}
\title{Tattersalls Sales in 2010}
\format{7796 rows, 12 columns}
\description{
Dataset of 12 Tattersalls sales in 2010
}
\section{tatts_2010}{

\itemize{
     \item dec_mares_2010 December Mares 2010 Sale
     \item dec_foals_2010 December Foals 2010 Sale
     \item dec_yearlings_2010 December Yearlings 2010 Sale
     \item autumn_hit_2010 Autumn Horses in Training 2010 Sale
     \item breeders_flat_2010 Breeders Flat 2010 Sale
     \item oct_bk3_2010 October Book 3 2010 Sale
     \item oct_bk2_2010 October Book 2 2010 Sale
     \item oct_bk1_2010 October Book 1 2010 Sale
     \item july_2010 July 2010 Sale
     \item guineas_2010 Breeze Up 2010 Sale
     \item craven_2010 Craven 2010 Sale
     \item feb_2010 February 2010 Sale
}
}
\keyword{dataset}

