% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{tatts_2011}
\alias{tatts_2011}
\title{Tattersalls Sales in 2011}
\format{6946 rows, 12 columns}
\description{
Dataset of 12 Tattersalls sales in 2011
}
\section{tatts_2011}{

\itemize{
     \item \strong{dec_mares_2011} December Mares 2011 Sale
     \item \strong{dec_foals_2011} December Foals 2011 Sale
     \item \strong{dec_yearlings_2011} December Yearlings 2011 Sale
     \item \strong{autumn_hit_2011} Autumn Horses in Training 2011 Sale
     \item \strong{breeders_flat_2011} Breeders Flat 2011 Sale
     \item \strong{oct_bk3_2011} October Book 3 2011 Sale
     \item \strong{oct_bk2_2011} October Book 2 2011 Sale
     \item \strong{oct_bk1_2011} October Book 1 2011 Sale
     \item \strong{july_2011} July 2011 Sale
     \item \strong{breeze_up_2011} Breeze Up 2011 Sale
     \item \strong{craven_2011} Craven 2011 Sale
     \item \strong{feb_2011} February 2011 Sale
}
}
\keyword{dataset}

