% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBloodstock.R
\docType{data}
\name{tatts_2013}
\alias{tatts_2013}
\title{Tattersalls Sales in 2013}
\format{7192 rows, 12 columns}
\description{
Dataset of 14 Tattersalls sales in 2013
}
\section{Sales}{

\itemize{
     \item dec_mares_2013 December Mares 2013 Sale
     \item dec_foals_2013 December Foals 2013 Sale
     \item dec_yearlings_2013 December Yearlings 2013 Sale
     \item autumn_hit_2013 Autumn Horses in Training 2013 Sale
     \item oct_bk4_2013 October Book 4 2013 Sale
     \item oct_bk3_2013 October Book 3 2013 Sale
     \item oct_bk2_2013 October Book 2 2013 Sale
     \item oct_bk1_2013 October Book 1 2013 Sale
     \item july_2013 July 2013 Sale
     \item guineas_hit_2013 Guineas Horses in Training 2013 Sale
     \item guineas_2013 Guineas 2013 Sale
     \item craven_2013 Craven 2013 Sale
     \item feb_2013 February 2013 Sale
}
}
\keyword{dataset}

