% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rBloodstock.R
\docType{data}
\name{tatts_2014}
\alias{tatts_2014}
\title{Tattersalls Sales in 2014}
\format{7429 rows, 12 columns}
\description{
Dataset of 13 Tattersalls sales in 2014
}
\section{tatts_2014}{

\itemize{
     \item dec_mares_2014 December Mares 2014 Sale
     \item dec_foals_2014 December Foals 2014 Sale
     \item dec_yearlings_2014 December Yearlings 2014 Sale
     \item autumn_hit_2014 Autumn Horses in Training 2014 Sale
     \item oct_bk4_2014 October Book 4 2014 Sale
     \item oct_bk3_2014 October Book 3 2014 Sale
     \item oct_bk2_2014 October Book 2 2014 Sale
     \item oct_bk1_2014 October Book 1 2014 Sale
     \item july_2014 July 2014 Sale
     \item guineas_hit_2014 Guineas Horses in Training 2014 Sale
     \item guineas_2014 Guineas 2014 Sale
     \item craven_2014 Craven 2014 Sale
     \item feb_2014 February 2014 Sale
}
}
\keyword{dataset}

