% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rBloodstock.R
\docType{data}
\name{tatts_2015}
\alias{tatts_2015}
\title{Tattersalls sales in 2015}
\format{5389 rows, 12 columns}
\description{
Dataset of Tattersalls sales in 2015
}
\section{tatts_2015}{

\itemize{
     \item autumn_hit Autumn Horses in Training Sale
     \item oct_bk4_2015 October Book 4 2015 Sale
     \item oct_bk3_2015 October Book 3 2015 Sale
     \item oct_bk2_2015 October Book 2 2015 Sale
     \item oct_bk1_2015 October Book 1 2015 Sale
     \item july_2015 July 2015 sale
     \item guineas_hit_2015 Guineas Horses in Training 2015 Sale
     \item guineas_2015 Guineas 2015 Sale
     \item craven_2015 Craven 2015 Sale
     \item feb_2015 February 2015 Sale
}
}
\keyword{dataset}

